Revert "media: use size_t for integer iterator to Vector::size()"
This reverts commit 8f86555e565f83e0ef570706320893c7c48a8950.

+++ include/media/AudioSystem.h
@@ -132,7 +132,7 @@ public:
                                       audio_stream_type_t stream = AUDIO_STREAM_DEFAULT);
 
     // return the number of input frames lost by HAL implementation, or 0 if the handle is invalid
-    static uint32_t getInputFramesLost(audio_io_handle_t ioHandle);
+    static size_t getInputFramesLost(audio_io_handle_t ioHandle);
 
     static int newAudioSessionId();
     static void acquireAudioSessionId(int audioSession);
